security

+++ vagrant/catalog/application.py
@@ -1,10 +1,29 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for
-from sqlalchemy import create_engine
+from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
+
+
+from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, CategoryItem
+from database_setup import Base, Category, CategoryItem, User
+
+from flask import session as login_session
+import random
+import string
+
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
 
 app = Flask(__name__)
 
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Catalog Menu Application"
+
+# Connect to Database and create database session
 engine = create_engine('sqlite:///catalog.db')
 Base.metadata.bind = engine
 
@@ -12,25 +31,161 @@ DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 
-# Fake Restaurants
-# category = {'name': 'The CRUDdy Crab', 'id': '1'}
-
-# categorys = [{'name': 'The CRUDdy Crab', 'id': '1'}, {'name':'Blue Burgers', 'id':'2'},{'name':'Taco Hut', 'id':'3'}]
-
-
-# Fake Menu Items
-# items = [ {'name':'Cheese Pizza', 'description':'made with fresh cheese', 'price':'$5.99','course' :'Entree', 'id':'1'}, {'name':'Chocolate Cake','description':'made with Dutch Chocolate', 'price':'$3.99', 'course':'Dessert','id':'2'},{'name':'Caesar Salad', 'description':'with fresh organic vegetables','price':'$5.99', 'course':'Entree','id':'3'},{'name':'Iced Tea', 'description':'with lemon','price':'$.99', 'course':'Beverage','id':'4'},{'name':'Spinach Dip', 'description':'creamy dip with fresh spinach','price':'$1.99', 'course':'Appetizer','id':'5'} ]
-# item =  {'name':'Cheese Pizza','description':'made with fresh cheese','price':'$5.99','course' :'Entree'}
-# items = []
+# Create anti-forgery state token
+@app.route('/login', methods=['GET', 'POST'])
+def showLogin():    
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)  
 
 
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']    
+    login_session['email'] = data['email']
+
+    # See if a user exists, if it doesn't make a new one
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+# User Helper Functions
+
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+@app.route('/gdisconnect')
+def gdisconnect():
+        # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+
+    if result['status'] == '200':
+        # Reset the user's sesson.
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
+# JSON APIs to view Catalog Information
 @app.route('/category/<int:category_id>/menu/JSON')
 def categoryMenuJSON(category_id):
     category = session.query(Restaurant).filter_by(id=category_id).one()
     items = session.query(MenuItem).filter_by(
         category_id=category_id).all()
     return jsonify(MenuItems=[i.serialize for i in items])
-
 
 #@app.route('/category/<int:category_id>/menu/<int:menu_id>/JSON')
 #def categoryItemJSON(category_id, menu_id):
@@ -48,7 +203,7 @@ def categoryMenuJSON(category_id):
 
 
 @app.route('/')
-@app.route('/categories/')
+@app.route('/category/')
 def showCategories():
     categories = session.query(Category).all()        
     return render_template('categories.html', categories=categories)
@@ -61,10 +216,13 @@ def showCategories():
 
 @app.route('/category/new/', methods=['GET', 'POST'])
 def newCategory():
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         newCategory = Category(name=request.form['name'])
-        session.add(newCategory)
+        session.add(newCategory)        
         session.commit()
+        flash('New Category %s Successfully Created' % newCategory.name)
         return redirect(url_for('showCategories'))
     else:
         return render_template('newcategory.html')
@@ -76,14 +234,17 @@ def newCategory():
 
 @app.route('/category/<int:category_id>/edit/', methods=['GET', 'POST'])
 def editCategory(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     editedCategory = session.query(
         Category).filter_by(id=category_id).one()
     if request.method == 'POST':
         if request.form['name']:
             editedCategory.name = request.form['name']
 
-        session.add(editedCategory)
+        session.add(editedCategory)        
         session.commit()
+        flash('Category Successfully Edited %s' % editedCategory.name)
         return redirect(url_for('showCategories'))
     else:
         return render_template('editcategory.html', category=editedCategory)
@@ -94,11 +255,14 @@ def editCategory(category_id):
 
 @app.route('/category/<int:category_id>/delete/', methods=['GET', 'POST'])
 def deleteCategory(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     categoryToDelete = session.query(
         Category).filter_by(id=category_id).one()
     if request.method == 'POST':
-        session.delete(categoryToDelete)
+        session.delete(categoryToDelete)        
         session.commit()
+        flash('%s Successfully Deleted' % categoryToDelete.name)
         return redirect(
             url_for('showCategories', category_id=category_id))
     else:
@@ -125,7 +289,9 @@ def showItems(category_id):
 
 
 @app.route('/category/<int:category_id>/item/new/', methods=['GET', 'POST'])
-def newCategoryItem(category_id):    
+def newCategoryItem(category_id):
+    if 'username' not in login_session:
+        return redirect('/login')  
     categories = session.query(Category).all()
     if request.method == 'POST':
         newItem = CategoryItem(name=request.form['name'], 
@@ -133,7 +299,7 @@ def newCategoryItem(category_id):
                             category_id=category_id)
         session.add(newItem)
         session.commit()
-
+        flash('New Category Item %s Item Successfully Created' % (newItem.name))            
         return redirect(url_for('showItems', category_id=category_id))
     else:
         return render_template('newcategoryitem.html', category_id=category_id, categories=categories)
@@ -147,6 +313,8 @@ def newCategoryItem(category_id):
 
 @app.route('/category/<int:category_id>/item/<int:item_id>/edit', methods=['GET', 'POST'])
 def editCategoryItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     categories = session.query(Category).all()
     editedItem = session.query(CategoryItem).filter_by(id=item_id).one()
     if request.method == 'POST':
@@ -156,7 +324,8 @@ def editCategoryItem(category_id, item_id):
             editedItem.description = request.form['description']
 
         session.add(editedItem)
-        session.commit()        
+        session.commit()
+        flash('Menu Item Successfully Edited')       
         return redirect(url_for('showItems', category_id=category_id))
     else:
         return render_template('editcategoryitem.html', category_id=category_id, item_id=item_id, item=editedItem, categories=categories)
@@ -168,10 +337,13 @@ def editCategoryItem(category_id, item_id):
 
 @app.route('/category/<int:category_id>/item/<int:item_id>/delete', methods=['GET', 'POST'])
 def deleteCategoryItem(category_id, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     itemToDelete = session.query(CategoryItem).filter_by(id=item_id).one()
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
+        flash('Menu Item Successfully Deleted')
         return redirect(url_for('showItems', category_id=category_id))
     else:
         return render_template('deletecategoryitem.html', item=itemToDelete)
@@ -224,5 +396,5 @@ def deleteCategoryItemDetail(category_id, item_id):
 
 if __name__ == '__main__':
     app.debug = True
-    #app.secret_key = 'super_secret_key'
+    app.secret_key = 'ThGRBgNnf-bf8cvRX8E5ftfm'
     app.run(host='0.0.0.0', port=5000)
